Improve the current git checkout prompt display
Using `git rev-parse --abbrev-ref HEAD` was pretty good, but only
returned the first ref for the given commit. That was almost always fine
but only showed "HEAD" if we had a detached branch checkout.

Using `git status` with this sed filtering means we'll still get the
current branch for normal branches like we did with ref-parse but when
detatched checkout of a branch on origin we'll see "HEAD detached at
origin/<whatever branch>"