Do some really fancy things for jasmine tests
The jasmine test filtering is real weird. You can't simply give it an
explict file to run: it has to be a pattern filter. Patterns are things
like "run all files with this filename" or "run all the specs in this
directory". Different projects will find tests differently based on
their spec setup. So I'm hacking to accomodate that.