Remove the ripgrep max-columns configuration
It kept limiting my results for long lines (as it should!) but it bugged
me for too long as I didn't remember that I'd configured that
limitation in ripgreprc.

If I really _need_ these limitations someday I'd rather find them in the
docs and add them to individual queries when needed.