Simplify git-begone finding branches
Remove the check to ignore our current branch and hardcoded special
branches. If our current branch is gone then things should get weird. If
special branches are gone then they are gone.

Switch to grep from rg: we don't need the speed of rg enough to break
the standard of grep in this case.