Allow providing :Rg with search parameters
Calling `Rg:<cr>` meant I was dropping into a ripgrep search with no
pattern. While I could then further filter with fzf it is not as
effective as searching using ripgrep in the first place.

Now the ctrl-g shortcut leaves me on a prompt to be able to either press
return and get the original no pattern behavior or provide a search
term to ripgrep.